refactor(memory_store): split content/compose.rs into compose/ directory
Splits the 1246-line compose.rs into four focused sub-modules:
- yaml.rs: scalar helpers, scan_fm_field, split_front_matter, source_tag
- chunk.rs: compose_chunk_file, rewrite_tags and email-specific helpers
- summary.rs: SummaryComposeInput, ComposedSummary, compose_summary_md, rewrite_summary_tags
- tests.rs: extracted #[cfg(test)] suite

All public symbols are re-exported from mod.rs to preserve the external API.
